pyproject.toml
[project]
name = "chat"
version = "0.1.0"
description = "Command-line real-time conversation app"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "openai>=1.0.0",
    "python-dotenv>=1.0.0",
    "prompt-toolkit>=3.0.0",
    "click>=8.1.0",
    "toml>=0.10.2",
    "tabulate>=0.9.0",
    "wcwidth>=0.2.13",
    "rich>=13.9.4",
]

[project.scripts]
chat-cli = "chat.cli:cli"
